+++ directProblemQuadratureMethod/array_utils.h
@@ -1,6 +1,6 @@
 #pragma once
 
-void WriteBasicArraysFile(std::vector<std::vector<std::vector<std::vector<std::complex<double>>>>> & a,
-	std::vector<std::vector<std::vector<std::complex<double>>>> & overline_a,
-	std::vector<std::vector<std::complex<double>>> & b);
+void WriteBasicArraysFile(std::vector<std::vector<std::vector<std::vector<std::complex<float>>>>> & a,
+	std::vector<std::vector<std::vector<std::complex<float>>>> & overline_a,
+	std::vector<std::vector<std::complex<float>>> & b);
